+++ src/bios/flib/src/io/ps2/mod.rs
@@ -1,5 +1,3 @@
-pub mod code;
-
 use core::arch::asm;
 
 pub fn send_data(data: u8) {
